+++ lib/redacted.dart
@@ -3,3 +3,4 @@ export 'package:redacted/src/redacted_configuration.dart';
 export 'package:redacted/src/redacted_widget.dart';
 export 'package:redacted/src/redacyed_style.dart';
 export 'package:redacted/src/unredacted_widget.dart';
+export 'package:redacted/src/redacted_hide_widget.dart';